Factor out to JVM-only
